Download RPC: do not reboot if fw upgrade fails due to server error ( OFW-3348 )
Supprts to trigger retry mechanism at tr069 agent.

Device is not retrying the Download when server is not responding, instead reboot happening.

Fix Description:
As per requirement, CPE has to retry firmware download till it reaches an hour of retry. Avoid rebooting the device in case of retry on the FwUpgradeManager

Validation performed:
1. Ensure DUT is online in ACS
2. Trigger Download RPC when image located server is down
3. CPE should retry to download firmware download
4. Take the packet capture
5. Check the TransferComplete message and got correct code:<cwmp:TransferComplete>
                <CommandKey/>
                <FaultStruct>
                    <FaultCode>
                        9010
                        </FaultCode>
                    <FaultString>
                        Download failure
                        </FaultString>
                    </FaultStruct>

Signed-off-by: Sarveswaran Nagarajan <sanagarajan.contractor@libertyglobal.com>