fix unit report reboot reason as Forced_Software_upgrade even though no firmware upgrade is triggered from ACS ( OFW-4496 F5685SI-1299 )
BUG DESCRIPTION:
Multiple reboot reason (Forced_Software_upgrade) is reported in telemetry.

Download RPC is triggered from acs and if the tftp server is not reachable then,
firmware download will be in "Retry" state and as part of retry mechanism the CPE will retry 60 times
in an interval of 1 min until it's successful and the CPE is setting the last
reboot reason as Forced_Software_upgrade all the time.

FIX DESCRIPTION:
As a fix, Modified the logic to set the last reboot reason only in firmware upgrade successful scenario.

TESTING PERFORMED:

1.Enable telemetry
2.Added firewall rules in CPE to make tftp server unreachable.
3.Upgrade the firmware from ACS.
4.Firmware download will be in Retry state.
5.Last reboot reason events won't be seen in telemetry.

Upstream Candidate: No.

Signed-off-by: Saravanakumar Karthikeyan <skarthikeyan.contractor@libertyglobal.com>